Fri, Dec 1, 2023, 11:41 PM +03:00

+++ meal/meal.py
@@ -9,7 +9,7 @@ def main():
 
 def convert(time):
      hour, minute = time.strip().split(":")
-     return int(hour) + int(minute) / 60 * 100
+     return float(hour) + float(minute) / 60 * 100
